Splitup mappings into separate groups for better organization
